+++ NEWS.md
@@ -1,3 +1,5 @@
+# gluedown (development version)
+
 # gluedown 1.0.2
 
 * Suggest `markdown` dependency and warn for `md_convert()` (#22).