+++ NEWS.md
@@ -1,5 +1,6 @@
-# gluedown (development version)
+# gluedown 1.0.3
 
+* Update `md_table()` tests for rvest 1.0.0 release.
 * Create `md_label()` to use with `md_reference()` (#25).
 * Try using clean `...` support in `md_link()` and `md_image()`
 * Use `goodpractices::gp()` to fix a few issues: